refactor: Update SequenceServerLayer component
